Add handleSpinWheel method and add SpinWheelConstants.java

+++ src/main/java/com/teamobi/mobiarmy2/model/User.java
@@ -208,7 +208,7 @@ public class User {
         return equip;
     }
 
-    public void updateItems(byte itemIndex, byte quantity) {
+    public void updateItems(byte itemIndex, int quantity) {
         this.items[itemIndex] += quantity;
         if (this.items[itemIndex] < 0) {
             this.items[itemIndex] = 0;
